+++ src/main/java/se/iths/corkdork/controller/ColorController.java
@@ -4,7 +4,7 @@ import se.iths.corkdork.entity.ColorEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import se.iths.corkdork.controller.service.ColorService;
+import se.iths.corkdork.service.ColorService;
 
 import java.util.Optional;
